Update TextInput, PasswordInput, and SelectInput components
